Simplify Pages
Every page basically stated with a ShellPage and everything else nested
under it.  Instead move the pertinent formatting into the common layout
and reap the rewards!
