+++ src/routes/(shell)/identity/groups/create/+page.svelte
@@ -7,7 +7,7 @@
 	import * as Identity from '$lib/openapi/identity';
 
 	import type { ShellPageSettings } from '$lib/layouts/types.ts';
-	import ShellPage from '$lib/layouts/ShellPage.svelte';
+	import ShellPageHeader from '$lib/layouts/ShellPageHeader.svelte';
 	import ShellMetadataSection from '$lib/layouts/ShellMetadataSection.svelte';
 	import ShellSection from '$lib/layouts/ShellSection.svelte';
 	import MultiSelect from '$lib/forms/MultiSelect.svelte';
@@ -59,62 +59,62 @@
 	);
 </script>
 
-<ShellPage {settings}>
-	<ShellMetadataSection metadata={resource.metadata} {names} bind:valid={metadataValid} />
+<ShellPageHeader {settings} />
 
-	<ShellSection title="Roles">
-		<MultiSelect
-			label="Select roles for group members."
-			hint="You must select at least one role."
-			options={roles}
-			value={resource.spec.roleIDs}
-			onValueChange={(e) => (resource.spec.roleIDs = e.value)}
-		>
-			{#snippet selected(value: string)}
-				{data.roles.find((x) => x.metadata.id == value)?.metadata.name}
-			{/snippet}
-		</MultiSelect>
-	</ShellSection>
+<ShellMetadataSection metadata={resource.metadata} {names} bind:valid={metadataValid} />
 
-	<ShellSection title="Users">
-		<MultiSelect
-			label="Select group members."
-			options={users}
-			value={resource.spec.userIDs}
-			onValueChange={(e) => (resource.spec.userIDs = e.value)}
-		>
-			{#snippet selected(value: string)}
-				{data.users.find((x) => x.metadata.id == value)?.spec.subject}
-			{/snippet}
-		</MultiSelect>
-	</ShellSection>
+<ShellSection title="Roles">
+	<MultiSelect
+		label="Select roles for group members."
+		hint="You must select at least one role."
+		options={roles}
+		value={resource.spec.roleIDs}
+		onValueChange={(e) => (resource.spec.roleIDs = e.value)}
+	>
+		{#snippet selected(value: string)}
+			{data.roles.find((x) => x.metadata.id == value)?.metadata.name}
+		{/snippet}
+	</MultiSelect>
+</ShellSection>
 
-	<ShellSection title="Service Accounts">
-		<MultiSelect
-			label="Select group members."
-			options={serviceAccounts}
-			value={resource.spec.serviceAccountIDs}
-			onValueChange={(e) => (resource.spec.serviceAccountIDs = e.value)}
-		>
-			{#snippet selected(value: string)}
-				{data.serviceAccounts.find((x) => x.metadata.id == value)?.metadata.name}
-			{/snippet}
-		</MultiSelect>
-	</ShellSection>
+<ShellSection title="Users">
+	<MultiSelect
+		label="Select group members."
+		options={users}
+		value={resource.spec.userIDs}
+		onValueChange={(e) => (resource.spec.userIDs = e.value)}
+	>
+		{#snippet selected(value: string)}
+			{data.users.find((x) => x.metadata.id == value)?.spec.subject}
+		{/snippet}
+	</MultiSelect>
+</ShellSection>
 
-	<div class="flex justify-between">
-		<Button
-			icon="mdi:cancel-bold"
-			label="Cancel"
-			class="preset-filled-surface-500"
-			href="/identity/groups"
-		/>
-		<Button
-			icon="mdi:tick"
-			label="Create"
-			class="preset-filled-primary-500"
-			clicked={submit}
-			disabled={!valid}
-		/>
-	</div>
-</ShellPage>
+<ShellSection title="Service Accounts">
+	<MultiSelect
+		label="Select group members."
+		options={serviceAccounts}
+		value={resource.spec.serviceAccountIDs}
+		onValueChange={(e) => (resource.spec.serviceAccountIDs = e.value)}
+	>
+		{#snippet selected(value: string)}
+			{data.serviceAccounts.find((x) => x.metadata.id == value)?.metadata.name}
+		{/snippet}
+	</MultiSelect>
+</ShellSection>
+
+<div class="flex justify-between">
+	<Button
+		icon="mdi:cancel-bold"
+		label="Cancel"
+		class="preset-filled-surface-500"
+		href="/identity/groups"
+	/>
+	<Button
+		icon="mdi:tick"
+		label="Create"
+		class="preset-filled-primary-500"
+		clicked={submit}
+		disabled={!valid}
+	/>
+</div>
