+++ src/routes/(shell)/identity/oauth2providers/create/+page.svelte
@@ -8,7 +8,7 @@
 	import * as Identity from '$lib/openapi/identity';
 
 	import type { ShellPageSettings } from '$lib/layouts/types.ts';
-	import ShellPage from '$lib/layouts/ShellPage.svelte';
+	import ShellPageHeader from '$lib/layouts/ShellPageHeader.svelte';
 	import ShellMetadataSection from '$lib/layouts/ShellMetadataSection.svelte';
 	import ShellSection from '$lib/layouts/ShellSection.svelte';
 	import TextInput from '$lib/forms/TextInput.svelte';
@@ -59,52 +59,51 @@
 		: '';
 </script>
 
-<ShellPage {settings}>
-	<ShellMetadataSection metadata={resource.metadata} {names} bind:valid={metadataValid} />
+<ShellPageHeader {settings} />
+<ShellMetadataSection metadata={resource.metadata} {names} bind:valid={metadataValid} />
 
-	<ShellSection title="OAuth2 Settings">
-		<label for="callback"
-			>Oauth2 callback address. This is used to configure your OIDC application with before
-			continuing with the following fields.</label
-		>
-		<Clipboard id="callback" value={callback} />
+<ShellSection title="OAuth2 Settings">
+	<label for="callback"
+		>Oauth2 callback address. This is used to configure your OIDC application with before continuing
+		with the following fields.</label
+	>
+	<Clipboard id="callback" value={callback} />
 
-		<TextInput
-			label="Issuer address"
-			validators={[Validation.stringSet]}
-			placeholder="https://identity.domain.com"
-			bind:value={resource.spec.issuer}
-			bind:valid={issuerValid}
-		/>
-		<TextInput
-			label="Client ID"
-			validators={[Validation.stringSet]}
-			placeholder="73458e95-1d2c-481b-81e8-7225fd089060"
-			bind:value={resource.spec.clientID}
-			bind:valid={clientIdValid}
-		/>
-		<TextInput
-			label="Client secret"
-			validators={[Validation.stringSet]}
-			placeholder="ooHovOvanogyisAvChuOvbyctoffOdloidKuAlsyemgosJias3twanechorjIdCo"
-			bind:value={resource.spec.clientSecret}
-			bind:valid={clientSecretValid}
-		/>
-	</ShellSection>
+	<TextInput
+		label="Issuer address"
+		validators={[Validation.stringSet]}
+		placeholder="https://identity.domain.com"
+		bind:value={resource.spec.issuer}
+		bind:valid={issuerValid}
+	/>
+	<TextInput
+		label="Client ID"
+		validators={[Validation.stringSet]}
+		placeholder="73458e95-1d2c-481b-81e8-7225fd089060"
+		bind:value={resource.spec.clientID}
+		bind:valid={clientIdValid}
+	/>
+	<TextInput
+		label="Client secret"
+		validators={[Validation.stringSet]}
+		placeholder="ooHovOvanogyisAvChuOvbyctoffOdloidKuAlsyemgosJias3twanechorjIdCo"
+		bind:value={resource.spec.clientSecret}
+		bind:valid={clientSecretValid}
+	/>
+</ShellSection>
 
-	<div class="flex justify-between">
-		<Button
-			icon="mdi:cancel-bold"
-			label="Cancel"
-			class="preset-filled-surface-500"
-			href="/identity/oauth2providers"
-		/>
-		<Button
-			icon="mdi:tick"
-			label="Create"
-			class="preset-filled-primary-500"
-			clicked={submit}
-			disabled={!valid}
-		/>
-	</div>
-</ShellPage>
+<div class="flex justify-between">
+	<Button
+		icon="mdi:cancel-bold"
+		label="Cancel"
+		class="preset-filled-surface-500"
+		href="/identity/oauth2providers"
+	/>
+	<Button
+		icon="mdi:tick"
+		label="Create"
+		class="preset-filled-primary-500"
+		clicked={submit}
+		disabled={!valid}
+	/>
+</div>
