+++ src/routes/(shell)/identity/serviceaccounts/+page.svelte
@@ -9,7 +9,7 @@
 	import * as Clients from '$lib/clients';
 
 	import type { ShellPageSettings } from '$lib/layouts/types.ts';
-	import ShellPage from '$lib/layouts/ShellPage.svelte';
+	import ShellPageHeader from '$lib/layouts/ShellPageHeader.svelte';
 	import ShellList from '$lib/layouts/ShellList.svelte';
 	import ShellListItem from '$lib/layouts/ShellListItem.svelte';
 	import ShellListItemHeader from '$lib/layouts/ShellListItemHeader.svelte';
@@ -45,46 +45,46 @@
 	}
 </script>
 
-<ShellPage {settings}>
+<ShellPageHeader {settings}>
 	{#snippet tools()}
 		<SubtleButton icon="mdi:add" label="Create" href="/identity/serviceaccounts/create" />
 	{/snippet}
+</ShellPageHeader>
 
-	<ShellList>
-		{#each data.serviceAccounts || [] as resource}
-			<ShellListItem>
-				{#snippet main()}
-					<ShellListItemHeader
-						metadata={resource.metadata}
-						href="/identity/serviceaccounts/view/{resource.metadata.id}"
-					/>
-				{/snippet}
+<ShellList>
+	{#each data.serviceAccounts || [] as resource}
+		<ShellListItem>
+			{#snippet main()}
+				<ShellListItemHeader
+					metadata={resource.metadata}
+					href="/identity/serviceaccounts/view/{resource.metadata.id}"
+				/>
+			{/snippet}
 
-				{#snippet badges()}
-					<ShellListItemBadges metadata={resource.metadata} />
-				{/snippet}
+			{#snippet badges()}
+				<ShellListItemBadges metadata={resource.metadata} />
+			{/snippet}
 
-				<ShellListItemMetadata metadata={resource.metadata}></ShellListItemMetadata>
+			<ShellListItemMetadata metadata={resource.metadata}></ShellListItemMetadata>
 
-				<ShellListItemMetadata>
-					<ShellMetadataItem
-						icon="mdi:key-outline"
-						label="Expiry"
-						value={resource.status.expiry.toUTCString()}
-					/>
-				</ShellListItemMetadata>
+			<ShellListItemMetadata>
+				<ShellMetadataItem
+					icon="mdi:key-outline"
+					label="Expiry"
+					value={resource.status.expiry.toUTCString()}
+				/>
+			</ShellListItemMetadata>
 
-				{#snippet trail()}
-					<ModalIcon
-						icon="mdi:trash-can-outline"
-						label="Delete"
-						title="Are you sure?"
-						confirm={() => confirm(resource.metadata.id)}
-					>
-						Removing service account "{resource.metadata.name}" may break automation.
-					</ModalIcon>
-				{/snippet}
-			</ShellListItem>
-		{/each}
-	</ShellList>
-</ShellPage>
+			{#snippet trail()}
+				<ModalIcon
+					icon="mdi:trash-can-outline"
+					label="Delete"
+					title="Are you sure?"
+					confirm={() => confirm(resource.metadata.id)}
+				>
+					Removing service account "{resource.metadata.name}" may break automation.
+				</ModalIcon>
+			{/snippet}
+		</ShellListItem>
+	{/each}
+</ShellList>
