+++ src/routes/(shell)/identity/users/create/+page.svelte
@@ -8,7 +8,7 @@
 	import * as Validation from '$lib/validation';
 
 	import type { ShellPageSettings } from '$lib/layouts/types.ts';
-	import ShellPage from '$lib/layouts/ShellPage.svelte';
+	import ShellPageHeader from '$lib/layouts/ShellPageHeader.svelte';
 	import ShellSection from '$lib/layouts/ShellSection.svelte';
 	import TextInput from '$lib/forms/TextInput.svelte';
 	import MultiSelect from '$lib/forms/MultiSelect.svelte';
@@ -51,44 +51,43 @@
 	let groups = $derived(data.groups.map((x) => ({ label: x.metadata.name, value: x.metadata.id })));
 </script>
 
-<ShellPage {settings}>
-	<ShellSection title="User Information">
-		<TextInput
-			bind:value={resource.spec.subject}
-			label="Email Address."
-			hint="Users canonical email address, use of email aliases will not work."
-			validators={[Validation.stringSet, (name: string) => Validation.unique(name, names)]}
-			bind:valid={usernameValid}
-		/>
-	</ShellSection>
+<ShellPageHeader {settings} />
+<ShellSection title="User Information">
+	<TextInput
+		bind:value={resource.spec.subject}
+		label="Email Address."
+		hint="Users canonical email address, use of email aliases will not work."
+		validators={[Validation.stringSet, (name: string) => Validation.unique(name, names)]}
+		bind:valid={usernameValid}
+	/>
+</ShellSection>
 
-	<ShellSection title="Access Control">
-		<MultiSelect
-			label="Select group access."
-			hint="Groups associate users with projects and grant them permissions to create, view, edit and delete."
-			options={groups}
-			value={resource.spec.groupIDs}
-			onValueChange={(e) => (resource.spec.groupIDs = e.value)}
-		>
-			{#snippet selected(value: string)}
-				{data.groups.find((x) => x.metadata.id == value)?.metadata.name}
-			{/snippet}
-		</MultiSelect>
-	</ShellSection>
+<ShellSection title="Access Control">
+	<MultiSelect
+		label="Select group access."
+		hint="Groups associate users with projects and grant them permissions to create, view, edit and delete."
+		options={groups}
+		value={resource.spec.groupIDs}
+		onValueChange={(e) => (resource.spec.groupIDs = e.value)}
+	>
+		{#snippet selected(value: string)}
+			{data.groups.find((x) => x.metadata.id == value)?.metadata.name}
+		{/snippet}
+	</MultiSelect>
+</ShellSection>
 
-	<div class="flex justify-between">
-		<Button
-			icon="mdi:cancel-bold"
-			label="Cancel"
-			class="preset-filled-surface-500"
-			href="/identity/users"
-		/>
-		<Button
-			icon="mdi:tick"
-			label="Create"
-			class="preset-filled-primary-500"
-			clicked={submit}
-			disabled={!valid}
-		/>
-	</div>
-</ShellPage>
+<div class="flex justify-between">
+	<Button
+		icon="mdi:cancel-bold"
+		label="Cancel"
+		class="preset-filled-surface-500"
+		href="/identity/users"
+	/>
+	<Button
+		icon="mdi:tick"
+		label="Create"
+		class="preset-filled-primary-500"
+		clicked={submit}
+		disabled={!valid}
+	/>
+</div>
