+++ src/routes/(shell)/kubernetes/clustermanagers/+page.svelte
@@ -10,7 +10,7 @@
 	import * as Kubernetes from '$lib/openapi/kubernetes';
 
 	import type { ShellPageSettings } from '$lib/layouts/types.ts';
-	import ShellPage from '$lib/layouts/ShellPage.svelte';
+	import ShellPageHeader from '$lib/layouts/ShellPageHeader.svelte';
 	import ShellList from '$lib/layouts/ShellList.svelte';
 	import ShellListItem from '$lib/layouts/ShellListItem.svelte';
 	import ShellListItemHeader from '$lib/layouts/ShellListItemHeader.svelte';
@@ -47,32 +47,31 @@
 	}
 </script>
 
-<ShellPage {settings}>
-	<ShellList>
-		{#each data.clustermanagers as resource}
-			<ShellListItem>
-				{#snippet main()}
-					<ShellListItemHeader metadata={resource.metadata} projects={data.projects} />
-				{/snippet}
+<ShellPageHeader {settings} />
+<ShellList>
+	{#each data.clustermanagers as resource}
+		<ShellListItem>
+			{#snippet main()}
+				<ShellListItemHeader metadata={resource.metadata} projects={data.projects} />
+			{/snippet}
 
-				{#snippet badges()}
-					<ShellListItemBadges metadata={resource.metadata} />
-				{/snippet}
+			{#snippet badges()}
+				<ShellListItemBadges metadata={resource.metadata} />
+			{/snippet}
 
-				<ShellListItemMetadata metadata={resource.metadata} />
+			<ShellListItemMetadata metadata={resource.metadata} />
 
-				{#snippet trail()}
-					<ModalIcon
-						icon="mdi:trash-can-outline"
-						label="Delete"
-						title="Are you sure?"
-						confirm={() => confirm(resource)}
-					>
-						Removing cluster manager "{resource.metadata.name}" will remove any clusters managed by
-						it.
-					</ModalIcon>
-				{/snippet}
-			</ShellListItem>
-		{/each}
-	</ShellList>
-</ShellPage>
+			{#snippet trail()}
+				<ModalIcon
+					icon="mdi:trash-can-outline"
+					label="Delete"
+					title="Are you sure?"
+					confirm={() => confirm(resource)}
+				>
+					Removing cluster manager "{resource.metadata.name}" will remove any clusters managed by
+					it.
+				</ModalIcon>
+			{/snippet}
+		</ShellListItem>
+	{/each}
+</ShellList>
