+++ src/routes/(shell)/kubernetes/clusters/view/[id]/+page.svelte
@@ -8,7 +8,7 @@
 	import * as RegionUtil from '$lib/regionutil';
 
 	import type { ShellPageSettings } from '$lib/layouts/types.ts';
-	import ShellPage from '$lib/layouts/ShellPage.svelte';
+	import ShellPageHeader from '$lib/layouts/ShellPageHeader.svelte';
 	import ShellViewHeader from '$lib/layouts/ShellViewHeader.svelte';
 	import ShellMetadataSection from '$lib/layouts/ShellMetadataSection.svelte';
 	import ShellSection from '$lib/layouts/ShellSection.svelte';
@@ -247,163 +247,160 @@
 	}
 </script>
 
-<ShellPage {settings}>
-	<ShellViewHeader metadata={cluster.metadata}>
-		{#snippet badges()}
-			<Badge icon={RegionUtil.icon(data.regions, cluster.spec.regionId)}>
-				{RegionUtil.name(data.regions, cluster.spec.regionId)}
-			</Badge>
-		{/snippet}
-	</ShellViewHeader>
+<ShellPageHeader {settings} />
+<ShellViewHeader metadata={cluster.metadata}>
+	{#snippet badges()}
+		<Badge icon={RegionUtil.icon(data.regions, cluster.spec.regionId)}>
+			{RegionUtil.name(data.regions, cluster.spec.regionId)}
+		</Badge>
+	{/snippet}
+</ShellViewHeader>
 
-	<Stepper steps={3} bind:step {valid} {complete}>
-		{#snippet content(index: number)}
-			{#if index === 0}
-				<h2 class="h2">Basic Cluster Setup</h2>
+<Stepper steps={3} bind:step {valid} {complete}>
+	{#snippet content(index: number)}
+		{#if index === 0}
+			<h2 class="h2">Basic Cluster Setup</h2>
 
-				<ShellMetadataSection
-					metadata={cluster.metadata}
-					names={data.names}
-					bind:valid={metadataValid}
-				/>
+			<ShellMetadataSection
+				metadata={cluster.metadata}
+				names={data.names}
+				bind:valid={metadataValid}
+			/>
 
-				<ShellSection title="Platform Configuration">
-					<Select
-						label="Choose a Kubernetes version."
-						hint="Kubernetes provides guarantees backward
+			<ShellSection title="Platform Configuration">
+				<Select
+					label="Choose a Kubernetes version."
+					hint="Kubernetes provides guarantees backward
                                                 compatibility so choosing the newest is usually the right choice as that provides a rich
                                                 feature set and enhanced security. Certain applications — e.g. Kubeflow —
 						may require a specific version."
-						bind:value={cluster.spec.version}
-					>
-						{#each versions as version}
-							<option value={version}>{version}</option>
-						{/each}
-					</Select>
-				</ShellSection>
-			{:else if step === 1}
-				<ResourceList
-					title="Workload Pool Configuration"
-					columns={3}
-					items={cluster.spec.workloadPools}
-					bind:active={workloadPoolActive}
-					valid={workloadPoolValidFull}
-					add={workloadPoolAdd}
-					remove={workloadPoolRemove}
+					bind:value={cluster.spec.version}
 				>
-					{#snippet description()}
-						<p>
-							Workload pools provide compute resouce for your cluster. You may have as many as
-							required for your workload. Each pool has a set of CPU, GPU and memory that can be
-							selected from a defined set of flavours. Workload pools support automatic scaling,
-							thus reducing overall operational cost when not in use.
-						</p>
-					{/snippet}
-
-					<!-- eslint-disable @typescript-eslint/no-unused-vars -->
-					{#snippet normal(pool: Kubernetes.KubernetesClusterWorkloadPool, index: number)}
-						<div class="h5 font-bold">{pool.name}</div>
-
-						<div>{replicasString(pool)}</div>
-
-						<Flavor flavor={lookupFlavor(pool.machine.flavorId)} />
-					{/snippet}
-
-					<!-- eslint-disable @typescript-eslint/no-unused-vars -->
-					{#snippet expanded(pool: Kubernetes.KubernetesClusterWorkloadPool, index: number)}
-						<KubernetesWorkloadPool
-							flavors={data.flavors}
-							bind:pool={cluster.spec.workloadPools[index]}
-							bind:valid={workloadPoolValid}
-						/>
-					{/snippet}
-				</ResourceList>
-			{:else if index === 2}
-				<h2 class="h2">Advanced Options</h2>
-
-				<ShellSection title="Auto Upgrade">
+					{#each versions as version}
+						<option value={version}>{version}</option>
+					{/each}
+				</Select>
+			</ShellSection>
+		{:else if step === 1}
+			<ResourceList
+				title="Workload Pool Configuration"
+				columns={3}
+				items={cluster.spec.workloadPools}
+				bind:active={workloadPoolActive}
+				valid={workloadPoolValidFull}
+				add={workloadPoolAdd}
+				remove={workloadPoolRemove}
+			>
+				{#snippet description()}
 					<p>
-						Kubernetes clusters are provisioned using pre-defined bundles of applications. These are
-						periodically updated to provide security updates, bug fixes and platorm stability. These
-						are enabled by default to protect you and mitigate any issues that may arise.
+						Workload pools provide compute resouce for your cluster. You may have as many as
+						required for your workload. Each pool has a set of CPU, GPU and memory that can be
+						selected from a defined set of flavours. Workload pools support automatic scaling, thus
+						reducing overall operational cost when not in use.
 					</p>
+				{/snippet}
 
+				<!-- eslint-disable @typescript-eslint/no-unused-vars -->
+				{#snippet normal(pool: Kubernetes.KubernetesClusterWorkloadPool, index: number)}
+					<div class="h5 font-bold">{pool.name}</div>
+
+					<div>{replicasString(pool)}</div>
+
+					<Flavor flavor={lookupFlavor(pool.machine.flavorId)} />
+				{/snippet}
+
+				<!-- eslint-disable @typescript-eslint/no-unused-vars -->
+				{#snippet expanded(pool: Kubernetes.KubernetesClusterWorkloadPool, index: number)}
+					<KubernetesWorkloadPool
+						flavors={data.flavors}
+						bind:pool={cluster.spec.workloadPools[index]}
+						bind:valid={workloadPoolValid}
+					/>
+				{/snippet}
+			</ResourceList>
+		{:else if index === 2}
+			<h2 class="h2">Advanced Options</h2>
+
+			<ShellSection title="Auto Upgrade">
+				<p>
+					Kubernetes clusters are provisioned using pre-defined bundles of applications. These are
+					periodically updated to provide security updates, bug fixes and platorm stability. These
+					are enabled by default to protect you and mitigate any issues that may arise.
+				</p>
+
+				<Switch
+					name="autoupgrade"
+					label="Enable auto-upgrade"
+					hint="Upgrades may still occur as application bundles reach end-of-life even if you choose to opt out."
+					initial={$state.snapshot(cluster.spec.autoUpgrade) === undefined ||
+						Boolean($state.snapshot(cluster.spec.autoUpgrade?.enabled))}
+					onCheckedChange={autoUpgradeChange}
+				/>
+
+				{#if cluster.spec.autoUpgrade?.enabled}
 					<Switch
-						name="autoupgrade"
-						label="Enable auto-upgrade"
-						hint="Upgrades may still occur as application bundles reach end-of-life even if you choose to opt out."
-						initial={$state.snapshot(cluster.spec.autoUpgrade) === undefined ||
-							Boolean($state.snapshot(cluster.spec.autoUpgrade?.enabled))}
-						onCheckedChange={autoUpgradeChange}
+						name="autoupgradeoverride"
+						label="Override auto-upgrade default time windows"
+						hint="Auto upgrades are scheduled Monday-Friday beween 00:00 and 07:00 UTC.  This provides a good level of support coverage, and upgrades occur outside of European business hours."
+						initial={Boolean($state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek))}
+						onCheckedChange={autoUpgradeOverideChange}
 					/>
 
-					{#if cluster.spec.autoUpgrade?.enabled}
-						<Switch
-							name="autoupgradeoverride"
-							label="Override auto-upgrade default time windows"
-							hint="Auto upgrades are scheduled Monday-Friday beween 00:00 and 07:00 UTC.  This provides a good level of support coverage, and upgrades occur outside of European business hours."
-							initial={Boolean($state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek))}
-							onCheckedChange={autoUpgradeOverideChange}
-						/>
-
-						{#if cluster.spec.autoUpgrade?.daysOfWeek}
-							<div class="grid grid-cols-[auto_auto_1fr] gap-4">
-								<TimeWindow
-									title="Sunday"
-									checked={Boolean($state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.sunday))}
-									start={$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.sunday?.start)}
-									end={$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.sunday?.end)}
-									onChange={autoUpgradeChangeSunday}
-								/>
-								<TimeWindow
-									title="Monday"
-									checked={Boolean($state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.monday))}
-									start={$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.monday?.start)}
-									end={$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.monday?.end)}
-									onChange={autoUpgradeChangeMonday}
-								/>
-								<TimeWindow
-									title="Tuesday"
-									checked={Boolean($state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.tuesday))}
-									start={$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.tuesday?.start)}
-									end={$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.tuesday?.end)}
-									onChange={autoUpgradeChangeTuesday}
-								/>
-								<TimeWindow
-									title="Wednesday"
-									checked={Boolean(
-										$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.wednesday)
-									)}
-									start={$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.wednesday?.start)}
-									end={$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.wednesday?.end)}
-									onChange={autoUpgradeChangeWednesday}
-								/>
-								<TimeWindow
-									title="Thursday"
-									checked={Boolean($state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.thursday))}
-									start={$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.thursday?.start)}
-									end={$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.thursday?.end)}
-									onChange={autoUpgradeChangeThursday}
-								/>
-								<TimeWindow
-									title="Friday"
-									checked={Boolean($state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.friday))}
-									start={$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.friday?.start)}
-									end={$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.friday?.end)}
-									onChange={autoUpgradeChangeFriday}
-								/>
-								<TimeWindow
-									title="Saturday"
-									checked={Boolean($state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.saturday))}
-									start={$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.saturday?.start)}
-									end={$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.saturday?.end)}
-									onChange={autoUpgradeChangeSaturday}
-								/>
-							</div>
-						{/if}
+					{#if cluster.spec.autoUpgrade?.daysOfWeek}
+						<div class="grid grid-cols-[auto_auto_1fr] gap-4">
+							<TimeWindow
+								title="Sunday"
+								checked={Boolean($state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.sunday))}
+								start={$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.sunday?.start)}
+								end={$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.sunday?.end)}
+								onChange={autoUpgradeChangeSunday}
+							/>
+							<TimeWindow
+								title="Monday"
+								checked={Boolean($state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.monday))}
+								start={$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.monday?.start)}
+								end={$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.monday?.end)}
+								onChange={autoUpgradeChangeMonday}
+							/>
+							<TimeWindow
+								title="Tuesday"
+								checked={Boolean($state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.tuesday))}
+								start={$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.tuesday?.start)}
+								end={$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.tuesday?.end)}
+								onChange={autoUpgradeChangeTuesday}
+							/>
+							<TimeWindow
+								title="Wednesday"
+								checked={Boolean($state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.wednesday))}
+								start={$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.wednesday?.start)}
+								end={$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.wednesday?.end)}
+								onChange={autoUpgradeChangeWednesday}
+							/>
+							<TimeWindow
+								title="Thursday"
+								checked={Boolean($state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.thursday))}
+								start={$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.thursday?.start)}
+								end={$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.thursday?.end)}
+								onChange={autoUpgradeChangeThursday}
+							/>
+							<TimeWindow
+								title="Friday"
+								checked={Boolean($state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.friday))}
+								start={$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.friday?.start)}
+								end={$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.friday?.end)}
+								onChange={autoUpgradeChangeFriday}
+							/>
+							<TimeWindow
+								title="Saturday"
+								checked={Boolean($state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.saturday))}
+								start={$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.saturday?.start)}
+								end={$state.snapshot(cluster.spec.autoUpgrade?.daysOfWeek?.saturday?.end)}
+								onChange={autoUpgradeChangeSaturday}
+							/>
+						</div>
 					{/if}
-				</ShellSection>
-			{/if}
-		{/snippet}
-	</Stepper>
-</ShellPage>
+				{/if}
+			</ShellSection>
+		{/if}
+	{/snippet}
+</Stepper>
