+++ src/routes/(shell)/regions/networks/+page.svelte
@@ -9,7 +9,7 @@
 	import * as RegionUtil from '$lib/regionutil';
 
 	import type { ShellPageSettings } from '$lib/layouts/types.ts';
-	import ShellPage from '$lib/layouts/ShellPage.svelte';
+	import ShellPageHeader from '$lib/layouts/ShellPageHeader.svelte';
 	import ShellList from '$lib/layouts/ShellList.svelte';
 	import ShellListItem from '$lib/layouts/ShellListItem.svelte';
 	import ShellListItemHeader from '$lib/layouts/ShellListItemHeader.svelte';
@@ -32,46 +32,41 @@
 	});
 </script>
 
-<ShellPage {settings}>
-	<ShellList>
-		{#each data.networks as resource}
-			<ShellListItem>
-				{#snippet main()}
-					<ShellListItemHeader metadata={resource.metadata} projects={data.projects} />
-				{/snippet}
+<ShellPageHeader {settings} />
+<ShellList>
+	{#each data.networks as resource}
+		<ShellListItem>
+			{#snippet main()}
+				<ShellListItemHeader metadata={resource.metadata} projects={data.projects} />
+			{/snippet}
 
-				{#snippet badges()}
-					<ShellListItemBadges metadata={resource.metadata}>
-						{#snippet extra()}
-							<Badge icon={RegionUtil.icon(data.regions, resource.spec.regionId)}>
-								{RegionUtil.name(data.regions, resource.spec.regionId)}
-							</Badge>
-						{/snippet}
-					</ShellListItemBadges>
-				{/snippet}
+			{#snippet badges()}
+				<ShellListItemBadges metadata={resource.metadata}>
+					{#snippet extra()}
+						<Badge icon={RegionUtil.icon(data.regions, resource.spec.regionId)}>
+							{RegionUtil.name(data.regions, resource.spec.regionId)}
+						</Badge>
+					{/snippet}
+				</ShellListItemBadges>
+			{/snippet}
 
-				<ShellListItemMetadata metadata={resource.metadata} />
+			<ShellListItemMetadata metadata={resource.metadata} />
 
-				<ShellListItemMetadata>
+			<ShellListItemMetadata>
+				<ShellMetadataItem icon="mdi:network-outline" label="Prefix" value={resource.spec.prefix} />
+				<ShellMetadataItem
+					icon="mdi:dns-outline"
+					label="DNS Nameservers"
+					value={resource.spec.dnsNameservers.join(', ')}
+				/>
+				{#if resource.spec.openstack?.vlanId}
 					<ShellMetadataItem
-						icon="mdi:network-outline"
-						label="Prefix"
-						value={resource.spec.prefix}
+						icon="mdi:nic"
+						label="VLAN ID"
+						value={resource.spec.openstack.vlanId.toString()}
 					/>
-					<ShellMetadataItem
-						icon="mdi:dns-outline"
-						label="DNS Nameservers"
-						value={resource.spec.dnsNameservers.join(', ')}
-					/>
-					{#if resource.spec.openstack?.vlanId}
-						<ShellMetadataItem
-							icon="mdi:nic"
-							label="VLAN ID"
-							value={resource.spec.openstack.vlanId.toString()}
-						/>
-					{/if}
-				</ShellListItemMetadata>
-			</ShellListItem>
-		{/each}
-	</ShellList>
-</ShellPage>
+				{/if}
+			</ShellListItemMetadata>
+		</ShellListItem>
+	{/each}
+</ShellList>
